# api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Admin API",
    "version": "1.0.0",
    "description": "Express Admin 系统API文档，提供完整的用户管理、角色权限、内容管理等功能"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/articles": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "获取文章列表",
        "description": "获取文章列表，支持分页和搜索",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词（标题、内容、摘要）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "文章状态",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "offline",
                "pending"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "文章分类",
            "schema": {
              "type": "string",
              "enum": [
                "article",
                "news",
                "announcement",
                "tutorial"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "每页数量",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "list": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "文章管理"
        ],
        "summary": "创建新文章",
        "description": "创建新文章",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "文章标题"
                  },
                  "content": {
                    "type": "string",
                    "description": "文章内容"
                  },
                  "summary": {
                    "type": "string",
                    "description": "文章摘要"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "article",
                      "news",
                      "announcement",
                      "tutorial"
                    ],
                    "default": "article"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "文章标签"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "offline",
                      "pending"
                    ],
                    "default": "draft"
                  },
                  "featured": {
                    "type": "boolean",
                    "default": false
                  },
                  "coverImage": {
                    "type": "string",
                    "description": "封面图片URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "文章管理"
        ],
        "summary": "获取文章详情",
        "description": "根据文章ID获取文章详细信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "文章管理"
        ],
        "summary": "更新文章",
        "description": "更新文章信息",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "文章标题"
                  },
                  "content": {
                    "type": "string",
                    "description": "文章内容"
                  },
                  "summary": {
                    "type": "string",
                    "description": "文章摘要"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "article",
                      "news",
                      "announcement",
                      "tutorial"
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "文章标签"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "draft",
                      "published",
                      "offline",
                      "pending"
                    ]
                  },
                  "featured": {
                    "type": "boolean"
                  },
                  "coverImage": {
                    "type": "string",
                    "description": "封面图片URL"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "文章管理"
        ],
        "summary": "删除文章",
        "description": "删除指定文章",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "文章ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}